applications/spam/serializers.py
from rest_framework import serializers

from applications.spam.models import Contact


class ContactSerializer(serializers.ModelSerializer):

    class Meta:
        model = Contact
        fields = '__all__'


